requirements.txt
Django==4.2.11
pandas==2.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
qrcode==7.4.2
openpyxl==3.1.2

# Dev/test
pytest==8.3.3
pytest-django==4.12.0
